Restructure project to combine storybook and src
