+++ .storybook/preview.ts
@@ -1,0 +1,28 @@
+import type { Preview } from "@storybook/react-vite";
+import { CssBaseline, ThemeProvider, createTheme } from "@mui/material";
+import { withThemeFromJSXProvider } from "@storybook/addon-themes";
+import { BldrThemeProvider } from "../src/BldrThemeProvider";
+import "@fontsource/roboto/300.css";
+import "@fontsource/roboto/400.css";
+import "@fontsource/roboto/500.css";
+import "@fontsource/roboto/700.css";
+
+export const decorators = [
+  withThemeFromJSXProvider({
+    Provider: BldrThemeProvider,
+    GlobalStyles: CssBaseline,
+  }),
+];
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+};
+
+export default preview;
