--- package/package.json
@@ -1,74 +1,0 @@
-{
-  "name": "@mybldr/ui",
-  "version": "1.1.1",
-  "description": "myBLDR Component Library",
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "rollup": "rollup -c --bundleConfigAsCjs",
-    "release": "release-it"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mybldr/mybldr-ui.git"
-  },
-  "keywords": [
-    "myBLDR",
-    "component",
-    "library"
-  ],
-  "author": "",
-  "bugs": {
-    "url": "https://github.com/mybldr/mybldr-ui/issues"
-  },
-  "homepage": "https://github.com/mybldr/mybldr-ui#readme",
-  "dependencies": {
-    "@mui/icons-material": "7.1.0",
-    "@mui/material": "7.1.0",
-    "rollup": "4.40.0"
-  },
-  "peerDependencies": {
-    "@emotion/react": "^11.14.0",
-    "@emotion/styled": "^11.14.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.3.1"
-  },
-  "devDependencies": {
-    "@release-it/conventional-changelog": "10.0.1",
-    "@rollup/plugin-commonjs": "28.0.3",
-    "@rollup/plugin-node-resolve": "16.0.1",
-    "@rollup/plugin-terser": "0.4.4",
-    "@rollup/plugin-typescript": "12.1.2",
-    "auto-changelog": "2.5.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.3.1",
-    "release-it": "19.0.1",
-    "rollup-plugin-dts": "6.2.1",
-    "rollup-plugin-peer-deps-external": "2.2.4",
-    "tslib": "2.8.1",
-    "typescript": "5.8.3"
-  },
-  "release-it": {
-    "git": {
-      "commitMessage": "chore: release v${version}",
-      "changelog": "npx auto-changelog --stdout --commit-limit false --unreleased --template https://raw.githubusercontent.com/release-it/release-it/main/templates/changelog-compact.hbs"
-    },
-    "github": {
-      "release": true
-    },
-    "npm": {
-      "publish": true
-    },
-    "hooks": {
-      "after:bump": "npx auto-changelog -p"
-    }
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "overrides": {
-    "conventional-changelog-conventionalcommits": "8.0.0"
-  }
-}
