+++ src/Autocomplete.stories.tsx
@@ -1,8 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { Autocomplete } from "@mybldr/ui";
+import { Autocomplete } from "./Autocomplete";
 
-// More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
   title: "Autocomplete",
   component: Autocomplete,
