--- storybook/.storybook/main.ts
@@ -1,19 +1,0 @@
-import type { StorybookConfig } from "@storybook/react-vite";
-
-const config: StorybookConfig = {
-  stories: ["../src/**/*.mdx", "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
-  addons: [
-    "@storybook/addon-essentials",
-    "@storybook/addon-onboarding",
-    "@chromatic-com/storybook",
-    "@storybook/experimental-addon-test",
-  ],
-  framework: {
-    name: "@storybook/react-vite",
-    options: {},
-  },
-  typescript: {
-    reactDocgen: false,
-  },
-};
-export default config;
