--- storybook/.storybook/manager-head.html
@@ -1,1 +1,0 @@
-<meta name="robots" content="noindex" />
